after(): Document new callback signature
Since jQuery version 1.10, `$.fn.after` invokes the supplied function
with two arguments: the numeric index of each element in the set and the
string markup describing each element's contents. Remove the second
argument from the documentation of the pre-1.10 signature and add a new
entry to thoroughly document the post-1.10 signature.