Build: Add `.editorconfig` file (#877)
The vast majority of XML files in the repo use two spaces today,
and even the files that contain tabs, do so only among many more
lines in the same file that are indented with spaces instead.

As such, for now, normalize toward two spaces so that only a few
files need to edited.

The exception is the JS and XSL files, of which the are only two
of each. In accordance with jQuery Style Guide, keep those as tabs.

```
$ git grep $'\t' | cut -d':' -f1 | sort | uniq
Gruntfile.js
entries2html.xsl
notes.xsl
resources/events.js

git grep $'  ' | cut -d':' -f1 | sort | uniq
…
```

Closes https://github.com/jquery/api.jquery.com/pull/877.

Co-authored-by: Aurelio De Rosa <aurelioderosa@gmail.com>