jQuery.globalEval: Document the document parameter
Since jQuery 3.5.0, jQuery.globalEval accepts an optional third parameter
accepting a document in which context the code will be evaluated.

This commit also adds the Document type and links it to one of the jQuery
signatures; so far it was using `type="document"` which linked to a non-existent
types section.

Ref jquery/jquery#4601