Build: Skip checking Twitter links in the hydra crawler
Twitter pages now do 302-redirects to themselves for users without a
specific cookie set which trips the crawler; avoid checking Twitter
links by abusing the `exclude_scheme_prefixes` option of the crawler.

Since the project only accepts options in a form of a configuration
file, we also need to clone the API repo to provide such a file.